+++ R/cohort.R
@@ -1,16 +1,14 @@
-#' Runs a cohort discrete event simulation
-#'
-#' @param models Models in the form of a list of \code{flexsurvreg} objects.
-#' @param newdata Data frame with individual attributes
-#' @param trans_mat Transition matrix
-#' @param start_times Entry times of individuals specified in \code{newdata}.
-#'   Can either be a single time that everyone enters at, or have as many values
-#'   as rows in \code{newdata}. If NULL then everyone enter at t=0
-#' @return A data frame with state entry times for each individual.
-#' @export
-#'
-#' @useDynLib rdes
-#' @importFrom Rcpp evalCpp
+# Runs a cohort discrete event simulation
+#
+# @param models Models in the form of a list of \code{flexsurvreg} objects.
+# @param newdata Data frame with individual attributes
+# @param trans_mat Transition matrix
+# @param start_times Entry times of individuals specified in \code{newdata}.
+#  Can either be a single time that everyone enters at, or have as many values
+#   as rows in \code{newdata}. If NULL then everyone enter at t=0
+# @return A data frame with state entry times for each individual.
+# @export
+# @importFrom Rcpp evalCpp
 cohort_simulation <- function(models, newdata, trans_mat, start_times=NULL) {
 
     # TODO Should I also provide a higher level function that handles things like