+++ NEWS.md
@@ -1,3 +1,10 @@
+# 1.1.0
+
+Two new features added:
+
+  - A plot of dynamically predicted state pathways is available in `plot_predicted_pathways`. See `README` for an example!
+  - Length of stay estimates are now provided by `length_of_stay`
+
 # 1.0.0
 
 Initial release. Contains functionality for estimating transition probabilities in `predict_transitions`, with a similar interface to `flexsurv::pmatrix.simfs`, with additional features as described in *README.md*.